+++ __tests__/parser.js
@@ -1,6 +1,5 @@
-const {parseTag, parseInputList} = require('../src/parser')
-const {invalidTag} = require('../src/errors')
-const {parseTagScenarios, parseInputListScenarios} = require('./__data__/parser')
+const {parseTag} = require('../src/parseTag')
+const {parseTagScenarios} = require('./__data__/parser')
 
 describe('parseTag',()=>{
   test.each(parseTagScenarios)(
@@ -10,13 +9,4 @@ describe('parseTag',()=>{
     expect(tag).toMatch(expected)
     expect(error).toBeUndefined()
   })
-})
-
-describe('parseInputList', ()=>{
-
-  test.each(parseInputListScenarios)('expect %s to be an array%%', (list, expected) => {
-    let result = parseInputList(list)
-    expect(result).toEqual(expected)
-  })
-
 })
