+++ __tests__/taggingStrategy.js
@@ -1,5 +1,4 @@
 const {taggingStrategy} = require('../src/taggingStrategy')
-const {parseTag, getInputList} = require('../src/parser')
 const {scenarios} = require('./__data__/taggingStrategy')
 
 describe.each(scenarios)('produce tags %s given %o', (expected, given) => {
